move supermirror calibration and normalization out of amor
